==> 1.scrapping_tableaux_phase_de_groupe.ipynb <==
import pandas as pd
import pickle
from string import ascii_uppercase as alphabet

all_tables = pd.read_html('https://en.wikipedia.org/wiki/2022_FIFA_World_Cup')

dict_table = {}
for letter, i in zip(alphabet, range(9,61,7)): # A=11, B=18, ...
    df = all_tables[i]
    df.rename(columns={df.columns[1]:'Team'}, inplace=True)
    df.pop('Qualification')
    dict_table[f'Group {letter}'] = df

dict_table.keys()

dict_table['Group F']

with open('dict_table', 'wb') as output:
    pickle.dump(dict_table, output)
