==> 2.scrapping_match(Qata2022_worldcup).ipynb <==
import pandas as pd
from bs4 import BeautifulSoup
import requests

def get_matches(y):
    site = f'https://en.wikipedia.org/wiki/{y}_FIFA_World_Cup'
    req = requests.get(site)
    content = req.text
    soup = BeautifulSoup(content, 'lxml')
    matches = soup.find_all('div', class_='footballbox')
    home = []
    score = []
    away = []
    for match in matches:
        home.append(match.find('th', class_='fhome').get_text())
        score.append(match.find('th', class_='fscore').get_text())
        away.append(match.find('th', class_='faway').get_text())

    dict_foot= {'home': home, 'score': score, 'away': away}
    football = pd.DataFrame(dict_foot)
    football['year'] = y
    return football

years = [ 1930,1934,1950,1954,1958,1962,1966,1970,1974,1978,1982,1986,1990,1994,1998,2002,2006, 2010, 2014, 2018]

# les anciennes match

worldcup_match = [get_matches(year) for year in years]
worldcup_match = pd.concat(worldcup_match, ignore_index=True)
worldcup_match.to_csv("donnees/worldcup_match.csv", index=False)

#les matches de coupe du monde 2022

Qatar2022 = get_matches(2022)
Qatar2022.to_csv('donnees/Qatar2022_match.csv', index=False)
