==> 3.cleaning_worldcup_match.ipynb <==
import pandas as pd

match = pd.read_csv('donnees/Qatar2022_match.csv')

match['score'] = match['score'].str.replace('[^\d–]', '', regex=True)
match['score']='-'
match['home'] = match['home'].str.strip() # clean blank spaces: Yugoslavia twice
match['away'] = match['away'].str.strip()

match

match.to_csv('donnees/Qatar2022_match_Clean.csv',index=False)
