==> 4.cleaning_Qatatr2022_match.ipynb <==
import pandas as pd

match_coupdumonde = pd.read_csv('donnees/worldcup_match.csv')

match_coupdumonde.drop_duplicates(inplace=True)
match_coupdumonde.sort_values('year', inplace=True)

# nettoyer les colonnes
match_coupdumonde['score'] = match_coupdumonde['score'].str.replace('[^\d–]', '', regex=True)
match_coupdumonde['home'] = match_coupdumonde['home'].str.strip()
match_coupdumonde['away'] = match_coupdumonde['away'].str.strip()

# creation de deux nouveau colonnes et la supression de la colone score
match_coupdumonde[['HomeGoals', 'AwayGoals']] = match_coupdumonde['score'].str.split('–', expand=True)
match_coupdumonde.drop('score', axis=1, inplace=True)

# renommer et changer les types du colonnes
match_coupdumonde.rename(columns={'home': 'HomeTeam', 'away': 'AwayTeam',
                                   'year':'Year'}, inplace=True)
match_coupdumonde = match_coupdumonde.astype({'HomeGoals': int, 'AwayGoals':int, 'Year': int})

# creation de le colonne "totalgoals"
match_coupdumonde['TotalGoals'] = match_coupdumonde['HomeGoals'] + match_coupdumonde['AwayGoals']


match_coupdumonde.to_csv('donnees/worldcup_match_Clean.csv',index=False)
