==> 5.predection_Qatar2022.ipynb <==
import pandas as pd
import pickle
import numpy as np
from sklearn.model_selection import train_test_split
from sklearn.svm import SVC
from sklearn.ensemble import RandomForestClassifier

match=pd.read_csv("data 2/Qatar2022_match_Clean.csv")
equipe = pd.read_csv("data 2/Qatar2022_paticipant_Clean.csv", sep=";")
classement = pd.read_csv("data 2/classement_fifa_2022_Clean.csv")
data_match = pd.read_csv("data 2/lesmatches_internationaux_Clean.csv", sep=";")
dict_table = pickle.load(open('data 2/dict_table','rb'))

classement.head()

data_match.head()

match.head()

equipe.head()

data_match_df = data_match.replace({"IR Iran": "Iran", "Korea Republic" : "South Korea"})
classement_df = classement.replace({"IR Iran": "Iran", "Korea Republic" : "South Korea"})

data_match_df['rank_difference'] = data_match_df['home_team_fifa_rank'] - data_match_df['away_team_fifa_rank']
data_match_df['average_rank'] = (data_match_df['home_team_fifa_rank'] + data_match_df['away_team_fifa_rank'])/2
data_match_df['point_difference'] = data_match_df['home_team_total_fifa_points'] - data_match_df['away_team_total_fifa_points']
data_match_df['is_stake'] = data_match_df['tournament'] != 'Friendly'
data_match_df['is_worldcup'] = 'FIFA World Cup' in data_match_df['tournament']
data_match_df['score_difference'] = data_match_df['home_team_score'] - data_match_df['away_team_score']
data_match_df['is_won'] = data_match_df['score_difference'] > 0 # Take draw as lost


A= data_match_df.loc[:,['average_rank', 'rank_difference', 'point_difference', 'is_stake', 'is_worldcup']]
B=data_match_df['is_won']
A_train, A_test, B_train, B_test = train_test_split(A, B, test_size=0.2, random_state=42)

A

B

Support Vector Machines
Random Forest

A_train

A_test

B_train

B_test

# Support Vector Machines

svc = SVC()
svc.fit(A_train, B_train) #pour les variables d'entrainemenet
svm_pred = svc.predict(A_test)
accurrancy= round(svc.score(A_test, B_test) , 2)
accurrancy

# Random Forest

random_forest = RandomForestClassifier(n_estimators=50)
random_forest.fit(A_train, B_train)  #pour les variables d'entrainemenet
Y_pred = random_forest.predict(A_test)
random_forest.score(A_train, B_train)
accurrancy = round(random_forest.score(A_test, B_test), 2)
accurrancy


classement_df = classement_df.loc[(classement_df['rank_date'] == classement_df['rank_date'].max()) & (classement_df['country_full'].isin(equipe['Team'].unique()))]

classement_coup_du_monde = classement_df.set_index(['country_full'])

country_win_prob = {}
for country in equipe['Team'].unique():
    country_win_prob[country] = list()

country_win_prob

# Group Stage Match Prediction

def predict_points(home, away):

    # creer et initailiser les lignes et les column porte `average_rank', 'rank_difference', 'point_difference', 'is_stake', 'is_worldcup
        ligne = pd.DataFrame(np.array([[np.nan, np.nan, np.nan, True, True]]), columns=A_train.columns)

        home_rank = classement_coup_du_monde.loc[home, 'rank']
        home_points = classement_coup_du_monde.loc[home, 'total_points']
        away_rank = classement_coup_du_monde.loc[away, 'rank']
        away_points = classement_coup_du_monde.loc[away, 'total_points']

        ligne['average_rank'] = (home_rank + away_rank) / 2
        ligne['rank_difference'] = home_rank - away_rank
        ligne['point_difference'] = home_points - away_points

        modele = svc.predict(ligne)

        awaypoints=0
        homepoints=0

        if modele == False:
            awaypoints = 3

        elif modele== True:
            homepoints= 3

        return homepoints,awaypoints

diviser les phases de la competition

phase_de_groupes = match[:48].copy()
huitième_de_finale = match[48:56].copy()
quarts_de_finale = match[56:60].copy()
demi_finale= match[60:62].copy()
final = match[62:].copy()

for group in dict_table:
    teams_in_group = dict_table[group]['Team'].values
    match_group = phase_de_groupes[phase_de_groupes['home'].isin(teams_in_group)]
    for index, row in match_group.iterrows():
        home, away = row['home'], row['away']
        if home=='United States'or away=='United States':
            points_home=0
            points_away=0
        else:
            points_home, points_away = predict_points(home, away)
        dict_table[group].loc[dict_table[group]['Team'] == home, 'Pts'] += points_home
        dict_table[group].loc[dict_table[group]['Team'] == away, 'Pts'] += points_away

    dict_table[group] = dict_table[group].sort_values('Pts', ascending=False).reset_index()
    dict_table[group] = dict_table[group][['Team', 'Pts']]



dict_table


group_winner={}
for group in dict_table:
    group_winner[group] = dict_table[group].loc[0, 'Team'],dict_table[group].loc[1, 'Team']

group_winner

for group in dict_table:
    group_winner = dict_table[group].loc[0, 'Team']
    runners_up = dict_table[group].loc[1, 'Team']
    huitième_de_finale.replace({f'Winners {group}':group_winner,
                                 f'Runners-up {group}':runners_up}, inplace=True)

huitième_de_finale['winner'] = '?'
huitième_de_finale

def predict_victory(home, away):

    # creer et initailiser les lignes et les column porte `average_rank', 'rank_difference', 'point_difference', 'is_stake', 'is_worldcup
    ligne = pd.DataFrame(np.array([[np.nan, np.nan, np.nan, True, True]]), columns=A_train.columns)

    home_rank = classement_coup_du_monde.loc[home, 'rank']
    home_points = classement_coup_du_monde.loc[home, 'total_points']
    away_rank = classement_coup_du_monde.loc[away, 'rank']
    away_points = classement_coup_du_monde.loc[away, 'total_points']

    ligne['average_rank'] = (home_rank + away_rank) / 2
    ligne['rank_difference'] = home_rank - away_rank
    ligne['point_difference'] = home_points - away_points

    modele = random_forest.predict(ligne)

    if modele == False:
        victory = 'away'

    elif modele== True:
        victory = 'home'

    return  victory

def get_winner(match):
    for index, row in match.iterrows():
        home, away = row['home'], row['away']
        winn = predict_victory(home, away)
        if winn=='home':
            winner = home
        else:
            winner = away
        match.loc[index, 'winner'] = winner
    return match

get_winner(huitième_de_finale)

def update_table(round_1,round_2):
    for index, row in round_1.iterrows():
        winner = round_1.loc[index, 'winner']
        match = round_1.loc[index, 'match']
        round_2.replace({f'Winners {match}':winner}, inplace=True)
    round_2['winner'] = '?'
    return round_2

update_table(huitième_de_finale, quarts_de_finale)

get_winner(quarts_de_finale)

update_table(quarts_de_finale, demi_finale)

get_winner(demi_finale)

update_table(demi_finale,final)

get_winner(final)

# Merci pour votre attention
